services/survey-backend/requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
blinker==1.9.0
cachecontrol==0.14.4
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudevents==1.12.0
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi-mail==1.6.1
firebase-admin==7.1.0
firebase-functions==0.5.0
flask==3.1.2
flask-cors==6.0.2
functions-framework==3.10.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-events==0.5.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
msgpack==1.1.2
packaging==25.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyjwt==2.10.1
pymongo==4.16.0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
uvicorn-worker==0.4.0
watchdog==6.0.0
werkzeug==3.1.5
